==> mssm_pipeline/2_iterative_WGCNA_setup.Rmd <==
---
title: "Iterative WGCNA Setup"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

rm(list=ls())

library(dplyr)
library(ggplot2)
library(nnet)

source("../utils/plot_theme.R")
```

```{r load_data}
# Load clean gene expression data
mssm.cleaned <- readRDS("clean_data/1_mssm_cleaning/mssm_cleaned.RDS")

# Load module information by Ben
modules <- read.table("raw_data/Modules_March2018.txt", sep="\t", header=T)
module.names <- unique(modules$Module)

# Brain Regions
brs <- c("FP", "IFG", "PHG", "STG")
```

Module-Level Iterative WGCNA Setup
----------------------------------

I generated `.tsv` files that can be used as inputs for the `iterativeWGNCA` program on the cluster. In this case, we wanted to try running `iterativeWGCNA` on every module that Ben created for each study.

```{r}
# Remove header information
mssm <- mssm.cleaned[,-(1:21)]

# Transpose matrices
mssm <- t(mssm)

# Add gene names as the first column
mssm <- cbind(rownames(mssm), mssm)

# Add patients to the column names
colnames(mssm) <- c("Gene", mssm.cleaned$Patient)

# Read in module information
for (module.index in 1:length(module.names)) {
  
  # Skip all non-MSSM tissues
  if (all(sapply(brs, function(x) length(grep(x, module.names[module.index])) == 0))) {
    next
  }
  # MSSM modules
  # Extract genes and make sure MSSM transcripts exist
  # Write tab-spaced input file for iterativeWGCNA
  genes <- subset(modules, Module==module.names[module.index])$GeneID
  genes <- intersect(rownames(mssm), genes)
  write.table(mssm[as.character(genes),], file=paste0("clean_data/2_iterative_WGCNA_setup/", module.names[module.index], ".tsv"), row.names=F, col.names=T, sep="\t", quote=F)
}
```

Cohort-Level Iterative WGCNA Setup
----------------------------------

In this case, I generate files required to run `iterativeWGCNA` on each cohort. This involves ignoring Ben's modules and generating our own.

```{r}
write.table(mssm, file="clean_data/2_iterative_WGCNA_setup/iterativeWGCNA.tsv", row.names=F, col.names=T, sep="\t", quote=F)

```
