==> mssm_pipeline/6a_submodule_annotation.Rmd <==
---
title: "Cell Type Annotations for Submodules in MSSM PHG"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

rm(list=ls())

library(dplyr)
library(ggplot2)
library(clusterProfiler)
library(org.Hs.eg.db)
library(RColorBrewer)

source("../utils/plot_theme.R")

# Number of marker genes to use
NUM.MARKERS = 100
```

Cell type markers in the brain are taken from McKenzie et al. (Supplemental File 1, DOI:10.1038/s41598-018-27293-5). The study developed three different metrics to rank the relation of genes to specific cell types (enrichment, expression, specificity). 

```{r load_data}
# Get module genes for MSSM PHG
modules.genes <- read.table("raw_data/Modules_March2018.txt", sep="\t", header=T, stringsAsFactors=F) %>% 
  subset(brainRegion == "PHG") %>%
  dplyr::select(Module=Module, Gene=GeneID, HGNC=external_gene_name)
modules <- unique(modules.genes$Module)
m.gene.sets <- lapply(modules, function(x) subset(modules.genes, Module==x)[,2:3])
names(m.gene.sets) <- modules

# Submodule genes for MSSM PHG
# We will split these up by the module
submodules.genes <- readRDS("clean_data/3_iterative_WGCNA_cleaning/mssm_submodule_sig_genes.RDS")
submodules.genes <- submodules.genes %>% dplyr::filter(grepl("PHG", Submodule))

# Names of submodules
submodules <- unique(submodules.genes$Submodule)

# List of data frames by modules
gene.sets <- lapply(submodules, function(x) subset(submodules.genes, Submodule==x)[,2:3])
names(gene.sets) <- submodules

# Load data generated by McKenzie et al. (Supplemental File 1)
# https://doi.org/10.1038/s41598-018-27293-5
hum.enrich <- read.csv("raw_data/McKenzie_2017_Brain_Cell_Specific_Markers_Human_Enrich.csv", header=T, stringsAsFactors=F)
hum.express <- read.csv("raw_data/McKenzie_2017_Brain_Cell_Specific_Markers_Human_Expression.csv", header=T, stringsAsFactors=F)
hum.specific <- read.csv("raw_data/McKenzie_2017_Brain_Cell_Specific_Markers_Human_Specificity.csv", header=T, stringsAsFactors=F)

# Convert to lists based on cell type
# The authors recommend using the top 50 genes as markers
cell.types <- union(union(hum.enrich$Celltype, hum.express$Celltype), unique(hum.specific$Celltype))

hum.enrich <- lapply(cell.types, function(x) (subset(hum.enrich, Celltype == x) %>% dplyr::select(HGNC=gene, Grand.Mean=grand_mean))[1:NUM.MARKERS,])
hum.express <- lapply(cell.types, function(x) (subset(hum.express, Celltype == x) %>% dplyr::select(HGNC=gene, Grand.Rank=grand_rank))[1:NUM.MARKERS,])
hum.specific <- lapply(cell.types, function(x) (subset(hum.specific, Celltype == x) %>% dplyr::select(HGNC=gene, Grand.Mean=grand_mean))[1:NUM.MARKERS,])

cell.types <- c("Astroctyes", "Endothelial Cells", "Microglia", "Neurons", "Oligodendrocytes")

names(hum.enrich) <- cell.types
names(hum.express) <- cell.types
names(hum.specific) <- cell.types
```

Mapping HGNC to Ensembl
-----------------------

Cell type specific markers from McKenzie et al. are provided as HGNC symbols. I used `clusterProfiler` to map these to Ensembl accessions. 

```{r message=FALSE, warning=FALSE}
# McKenzie data set is using HGNC symbols
# Map to ENSEMBL symbols
hum.enrich <- lapply(cell.types, function(x) {
  gene.map <- bitr(hum.enrich[[x]]$HGNC, fromType="SYMBOL", toType="ENSEMBL", OrgDb=org.Hs.eg.db)
  hum.enrich[[x]] <- merge(hum.enrich[[x]], gene.map, by.x="HGNC", by.y="SYMBOL") %>% dplyr::arrange(desc(Grand.Mean))
})
names(hum.enrich) <- cell.types

hum.express <- lapply(cell.types, function(x) {
  gene.map <- bitr(hum.express[[x]]$HGNC, fromType="SYMBOL", toType="ENSEMBL", OrgDb=org.Hs.eg.db)
  hum.express[[x]] <- merge(hum.express[[x]], gene.map, by.x="HGNC", by.y="SYMBOL") %>% dplyr::arrange(desc(Grand.Rank))
})
names(hum.express) <- cell.types

hum.specific <- lapply(cell.types, function(x) {
  gene.map <- bitr(hum.specific[[x]]$HGNC, fromType="SYMBOL", toType="ENSEMBL", OrgDb=org.Hs.eg.db)
  hum.specific[[x]] <- merge(hum.specific[[x]], gene.map, by.x="HGNC", by.y="SYMBOL") %>% dplyr::arrange(desc(Grand.Mean))
})
names(hum.specific) <- cell.types
```

Cell Enrichment Score Ranks
---------------------------

I used the top markers from the cell enrichment score to annotate the submodules in the MSSM PHG cohort. I performed a chi-square test to compare the frequency of cell type markers in any given submodule to the total frequency of markers used. I performed a similar analysis for the modules. 

The cell enrichment for a given gene is the log fold change when comparing the expression of the gene in the given cell type with the expression of the gene in all other cell types.

```{r warning=FALSE}
# Submodules
enrich.set <- lapply(submodules, function(s) {
  sapply(cell.types, function(x) hum.enrich[[x]]$ENSEMBL %in% gene.sets[[s]]$Gene %>% which %>% length, USE.NAMES=T)
})
names(enrich.set) <- submodules
enrich.set <- do.call("rbind", enrich.set)

chi.sq.p <- sapply(submodules, function(s) chisq.test(rbind(enrich.set[s,], sapply(hum.enrich, nrow)))$p.value)
chi.sq.p <- p.adjust(chi.sq.p, method="bonferroni")

plot.data <- as.data.frame(stack(enrich.set))[,c(1,2,4)]
colnames(plot.data) <- c("Submodule", "Cell.Type", "Number.Genes")
p <- ggplot() +
  geom_bar(data=plot.data, aes(x=Submodule, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  geom_text(aes(x=submodules, y=rowSums(enrich.set), label=ifelse(chi.sq.p < 0.05, "*", "")), size=5, vjust=-0.1) +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("Submodule") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(axis.text.x=element_text(angle=60, hjust=1))
print(p)
ggsave(filename="results/6_submodule_annotation/hum.enrich.annot.svg", plot=p, width=8, height=7.5)

# Modules
enrich.set <- lapply(modules, function(s) {
  sapply(cell.types, function(x) hum.enrich[[x]]$ENSEMBL %in% m.gene.sets[[s]]$Gene %>% which %>% length, USE.NAMES=T)
})
names(enrich.set) <- modules
enrich.set <- do.call("rbind", enrich.set)

chi.sq.p <- sapply(modules, function(s) chisq.test(rbind(enrich.set[s,], sapply(hum.enrich, nrow)))$p.value)
chi.sq.p <- p.adjust(chi.sq.p, method="bonferroni")

plot.data <- as.data.frame(stack(enrich.set))[,c(1,2,4)]
colnames(plot.data) <- c("Module", "Cell.Type", "Number.Genes")
p <- ggplot() +
  geom_bar(data=plot.data, aes(x=Module, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  geom_text(aes(x=modules, y=rowSums(enrich.set), label=ifelse(chi.sq.p < 0.05, "*", "")), size=5, vjust=-0.1) +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("Module") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(axis.text.x=element_text(angle=60, hjust=1))
print(p)
ggsave(filename="results/6_submodule_annotation/m.hum.enrich.annot.svg", plot=p, width=5, height=7.5)
```

Cell Expression Score Ranks
---------------------------

I used the top markers from the cell expression score to annotate the submodules in the MSSM PHG cohort. I performed a chi-square test to compare the frequency of cell type markers in any given submodule to the total frequency of markers used. I performed a similar analysis for the modules. 

The cell expression for a given gene is the absolute expression value for the given gene in a given cell type.

```{r warning=FALSE}
# Submodules
express.set <- lapply(submodules, function(s) {
  sapply(cell.types, function(x) hum.express[[x]]$ENSEMBL %in% gene.sets[[s]]$Gene %>% which %>% length, USE.NAMES=T)
})
names(express.set) <- submodules
express.set <- do.call("rbind", express.set)

chi.sq.p <- sapply(submodules, function(s) chisq.test(rbind(express.set[s,], sapply(hum.express, nrow)))$p.value)
chi.sq.p <- p.adjust(chi.sq.p, method="bonferroni")

plot.data <- as.data.frame(stack(express.set))[,c(1,2,4)]
colnames(plot.data) <- c("Submodule", "Cell.Type", "Number.Genes")
p <- ggplot() +
  geom_bar(data=plot.data, aes(x=Submodule, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  geom_text(aes(x=submodules, y=rowSums(express.set), label=ifelse(chi.sq.p < 0.05, "*", "")), size=5, vjust=-0.1) +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("Submodule") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(axis.text.x=element_text(angle=60, hjust=1))
print(p)
ggsave(filename="results/6_submodule_annotation/hum.express.annot.svg", plot=p, width=8, height=7.5)

# Modules
express.set <- lapply(modules, function(s) {
  sapply(cell.types, function(x) hum.express[[x]]$ENSEMBL %in% m.gene.sets[[s]]$Gene %>% which %>% length, USE.NAMES=T)
})
names(express.set) <- modules
express.set <- do.call("rbind", express.set)

chi.sq.p <- sapply(modules, function(s) chisq.test(rbind(express.set[s,], sapply(hum.express, nrow)))$p.value)
chi.sq.p <- p.adjust(chi.sq.p, method="bonferroni")

plot.data <- as.data.frame(stack(express.set))[,c(1,2,4)]
colnames(plot.data) <- c("Module", "Cell.Type", "Number.Genes")
p <- ggplot() +
  geom_bar(data=plot.data, aes(x=Module, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  geom_text(aes(x=modules, y=rowSums(express.set), label=ifelse(chi.sq.p < 0.05, "*", "")), size=5, vjust=-0.1) +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("Module") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(axis.text.x=element_text(angle=60, hjust=1))
print(p)
ggsave(filename="results/6_submodule_annotation/m.hum.express.annot.svg", plot=p, width=5, height=7.5)
```

Cell Specificity Score Ranks
---------------------------

I used the top markers from the cell specificity score to annotate the submodules in the MSSM PHG cohort. I performed a chi-square test to compare the frequency of cell type markers in any given submodule to the total frequency of markers used. I performed a similar analysis for the modules. 

The cell specificity for a given gene is the minimum log fold change when pairwise comparing the expression of the gene in the given cell type with the expression of the gene in all other cell types.

```{r warning=FALSE}
# Submodules
specific.set <- lapply(submodules, function(s) {
  sapply(cell.types, function(x) hum.specific[[x]]$ENSEMBL %in% gene.sets[[s]]$Gene %>% which %>% length, USE.NAMES=T)
})
names(specific.set) <- submodules
specific.set <- do.call("rbind", specific.set)

chi.sq.p <- sapply(submodules, function(s) chisq.test(rbind(specific.set[s,], sapply(hum.specific, nrow)))$p.value)
chi.sq.p <- p.adjust(chi.sq.p, method="bonferroni")

plot.data.1 <- as.data.frame(stack(specific.set))[,c(1,2,4)]
colnames(plot.data.1) <- c("Submodule", "Cell.Type", "Number.Genes")
p <- ggplot() +
  geom_bar(data=plot.data.1, aes(x=Submodule, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  geom_text(aes(x=submodules, y=rowSums(specific.set), label=ifelse(chi.sq.p < 0.05, "*", "")), size=5, vjust=-0.1) +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("Submodule") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(axis.text.x=element_text(angle=60, hjust=1))
print(p)
ggsave(filename="results/6_submodule_annotation/hum.specific.annot.svg", plot=p, width=8, height=7.5)

# Modules
specific.set <- lapply(modules, function(s) {
  sapply(cell.types, function(x) hum.specific[[x]]$ENSEMBL %in% m.gene.sets[[s]]$Gene %>% which %>% length, USE.NAMES=T)
})
names(specific.set) <- modules
specific.set <- do.call("rbind", specific.set)

chi.sq.p <- sapply(modules, function(s) chisq.test(rbind(specific.set[s,], sapply(hum.specific, nrow)))$p.value)
chi.sq.p <- p.adjust(chi.sq.p, method="bonferroni")

plot.data.2 <- as.data.frame(stack(specific.set))[,c(1,2,4)]
colnames(plot.data.2) <- c("Module", "Cell.Type", "Number.Genes")
p <- ggplot() +
  geom_bar(data=plot.data.2, aes(x=Module, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  geom_text(aes(x=modules, y=rowSums(specific.set), label=ifelse(chi.sq.p < 0.05, "*", "")), size=5, vjust=-0.1) +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("Module") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(axis.text.x=element_text(angle=60, hjust=1))
print(p)
ggsave(filename="results/6_submodule_annotation/m.hum.specific.annot.svg", plot=p, width=5, height=7.5)
```

Combined Graph for Specificity
------------------------------

I found that the cell specificity markers were the best for demonstrating the cell type specificity of the submodules. I created a graphic that combines the modules and submodules into one image.

```{r}
plot.data <- rbind(
  plot.data.2 %>%
    dplyr::mutate(Scope="Module"),
  plot.data.1 %>% 
    dplyr::select(Module=Submodule, everything()) %>%
    dplyr::mutate(Scope="Submodule")
)
plot.data$Module <- factor(plot.data$Module, levels=c(modules, submodules))

p <- ggplot() +
  geom_bar(data=plot.data, aes(x=Module, y=Number.Genes, fill=Cell.Type), stat="identity", color="black") +
  guides(fill=guide_legend(title="Cell Type")) +
  scale_fill_brewer(palette="Set3") +
  xlab("") + ylab("Number of Marker Genes") +
  basic_theme_big +
  theme(
    axis.text.x=element_text(angle=60, hjust=1),
    legend.position="bottom"
  )
print(p)
ggsave(filename="results/6_submodule_annotation/cellspecific.annot.svg", plot=p, width=13, height=8)
```
